rename some sidebar files
